+++ src/main/java/org/jabref/model/entry/types/EntryType.java
@@ -4,10 +4,11 @@ public interface EntryType {
 
     /**
      * Returns the tag name of the entry type.
-     *
-     * @return tag name of the entry type.
      */
     String getName();
 
+    /**
+     * Returns the name presented in the UI
+     */
     String getDisplayName();
 }